@@ -1,14 +1,12 @@
 ﻿using Microsoft.Extensions.Logging;
-using System;
 
-namespace Console.ExampleFormatters.Custom
+namespace Console.ExampleFormatters.Custom;
+
+public static class ConsoleLoggerExtensions
 {
-    public static class ConsoleLoggerExtensions
-    {
-        public static ILoggingBuilder AddCustomFormatter(
-            this ILoggingBuilder builder,
-            Action<CustomOptions> configure) =>
-            builder.AddConsole(options => options.FormatterName = "customName")
-                .AddConsoleFormatter<CustomFormatter, CustomOptions>(configure);
-    }
+    public static ILoggingBuilder AddCustomFormatter(
+        this ILoggingBuilder builder,
+        Action<CustomOptions> configure) =>
+        builder.AddConsole(options => options.FormatterName = "customName")
+            .AddConsoleFormatter<CustomFormatter, CustomOptions>(configure);
 }
